+++ libs/community/tests/integration_tests/chat_models/test_moonshot.py
@@ -4,7 +4,7 @@ from typing import Type, cast
 
 import pytest
 from langchain_core.language_models import BaseChatModel
-from langchain_standard_tests.integration_tests import ChatModelIntegrationTests
+from langchain_tests.integration_tests import ChatModelIntegrationTests
 from pydantic import SecretStr
 
 from langchain_community.chat_models.moonshot import MoonshotChat
