+++ libs/core/tests/unit_tests/indexing/test_in_memory_indexer.py
@@ -3,7 +3,7 @@
 from collections.abc import AsyncGenerator, Generator
 
 import pytest
-from langchain_standard_tests.integration_tests.indexer import (
+from langchain_tests.integration_tests.indexer import (
     AsyncDocumentIndexTestSuite,
     DocumentIndexerTestSuite,
 )
