+++ libs/partners/fireworks/tests/unit_tests/test_standard.py
@@ -3,7 +3,7 @@
 from typing import Tuple, Type
 
 from langchain_core.language_models import BaseChatModel
-from langchain_standard_tests.unit_tests import (  # type: ignore[import-not-found]
+from langchain_tests.unit_tests import (  # type: ignore[import-not-found]
     ChatModelUnitTests,  # type: ignore[import-not-found]
 )
