+++ libs/partners/groq/tests/integration_tests/test_chat_models.py
@@ -395,7 +395,7 @@ def test_json_mode_structured_output() -> None:
 
 
 def test_tool_calling_no_arguments() -> None:
-    # Note: this is a variant of a test in langchain_standard_tests
+    # Note: this is a variant of a test in langchain_tests
     # that as of 2024-08-19 fails with "Failed to call a function. Please
     # adjust your prompt." when `tool_choice="any"` is specified, but
     # passes when `tool_choice` is not specified.
