+++ libs/partners/huggingface/tests/integration_tests/test_standard.py
@@ -4,7 +4,7 @@ from typing import Type
 
 import pytest
 from langchain_core.language_models import BaseChatModel
-from langchain_standard_tests.integration_tests import ChatModelIntegrationTests
+from langchain_tests.integration_tests import ChatModelIntegrationTests
 
 from langchain_huggingface import ChatHuggingFace, HuggingFaceEndpoint
