+++ libs/partners/ollama/tests/integration_tests/test_embeddings.py
@@ -2,7 +2,7 @@
 
 from typing import Type
 
-from langchain_standard_tests.integration_tests import EmbeddingsIntegrationTests
+from langchain_tests.integration_tests import EmbeddingsIntegrationTests
 
 from langchain_ollama.embeddings import OllamaEmbeddings
