+++ libs/partners/openai/tests/integration_tests/chat_models/test_base.py
@@ -22,10 +22,8 @@ from langchain_core.messages import (
 )
 from langchain_core.outputs import ChatGeneration, ChatResult, LLMResult
 from langchain_core.prompts import ChatPromptTemplate
-from langchain_standard_tests.integration_tests.chat_models import (
-    _validate_tool_call_message,
-)
-from langchain_standard_tests.integration_tests.chat_models import (
+from langchain_tests.integration_tests.chat_models import _validate_tool_call_message
+from langchain_tests.integration_tests.chat_models import (
     magic_function as invalid_magic_function,
 )
 from pydantic import BaseModel, Field
@@ -1022,7 +1020,8 @@ def test_audio_input_modality() -> None:
 
 
 def test_prediction_tokens() -> None:
-    code = dedent("""
+    code = dedent(
+        """
     /// <summary>
     /// Represents a user with a first name, last name, and username.
     /// </summary>
@@ -1043,7 +1042,8 @@ def test_prediction_tokens() -> None:
         /// </summary>
         public string Username { get; set; }
     }
-    """)
+    """
+    )
 
     llm = ChatOpenAI(model="gpt-4o")
     query = (
