+++ libs/partners/openai/tests/unit_tests/chat_models/test_base_standard.py
@@ -3,7 +3,7 @@
 from typing import Tuple, Type
 
 from langchain_core.language_models import BaseChatModel
-from langchain_standard_tests.unit_tests import ChatModelUnitTests
+from langchain_tests.unit_tests import ChatModelUnitTests
 
 from langchain_openai import ChatOpenAI
