+++ libs/partners/xai/tests/integration_tests/test_chat_models_standard.py
@@ -5,7 +5,7 @@ from typing import Optional, Type
 import pytest  # type: ignore[import-not-found]
 from langchain_core.language_models import BaseChatModel
 from langchain_core.rate_limiters import InMemoryRateLimiter
-from langchain_standard_tests.integration_tests import (  # type: ignore[import-not-found]
+from langchain_tests.integration_tests import (  # type: ignore[import-not-found]
     ChatModelIntegrationTests,  # type: ignore[import-not-found]
 )
