+++ libs/standard-tests/Makefile
@@ -25,7 +25,7 @@ PYTHON_FILES=.
 MYPY_CACHE=.mypy_cache
 lint format: PYTHON_FILES=.
 lint_diff format_diff: PYTHON_FILES=$(shell git diff --relative=libs/standard-tests --name-only --diff-filter=d master | grep -E '\.py$$|\.ipynb$$')
-lint_package: PYTHON_FILES=langchain_standard_tests
+lint_package: PYTHON_FILES=langchain_tests
 lint_tests: PYTHON_FILES=tests
 lint_tests: MYPY_CACHE=.mypy_cache_test
 
@@ -44,7 +44,7 @@ spell_check:
 spell_fix:
 	poetry run codespell --toml pyproject.toml -w
 
-check_imports: $(shell find langchain_standard_tests -name '*.py')
+check_imports: $(shell find langchain_tests -name '*.py')
 	poetry run python ./scripts/check_imports.py $^
 
 ######################
