+++ libs/standard-tests/langchain_tests/integration_tests/base_store.py
@@ -4,7 +4,7 @@ from typing import AsyncGenerator, Generator, Generic, Tuple, TypeVar
 import pytest
 from langchain_core.stores import BaseStore
 
-from langchain_standard_tests.base import BaseStandardTests
+from langchain_tests.base import BaseStandardTests
 
 V = TypeVar("V")
