+++ libs/standard-tests/langchain_tests/integration_tests/cache.py
@@ -4,7 +4,7 @@ import pytest
 from langchain_core.caches import BaseCache
 from langchain_core.outputs import Generation
 
-from langchain_standard_tests.base import BaseStandardTests
+from langchain_tests.base import BaseStandardTests
 
 
 class SyncCacheTestSuite(BaseStandardTests):
