+++ libs/standard-tests/langchain_tests/integration_tests/tools.py
@@ -1,7 +1,7 @@
 from langchain_core.messages import ToolCall
 from langchain_core.tools import BaseTool
 
-from langchain_standard_tests.unit_tests.tools import ToolsTests
+from langchain_tests.unit_tests.tools import ToolsTests
 
 
 class ToolsIntegrationTests(ToolsTests):
