+++ libs/standard-tests/tests/unit_tests/test_in_memory_base_store.py
@@ -5,7 +5,7 @@ from typing import Tuple
 import pytest
 from langchain_core.stores import InMemoryStore
 
-from langchain_standard_tests.integration_tests.base_store import (
+from langchain_tests.integration_tests.base_store import (
     BaseStoreAsyncTests,
     BaseStoreSyncTests,
 )
