+++ libs/standard-tests/tests/unit_tests/test_in_memory_vectorstore.py
@@ -4,7 +4,7 @@ from langchain_core.vectorstores import (
     VectorStore,
 )
 
-from langchain_standard_tests.integration_tests.vectorstores import (
+from langchain_tests.integration_tests.vectorstores import (
     AsyncReadWriteTestSuite,
     ReadWriteTestSuite,
 )
